projects/marketplace/backend/requirements.txt
# MyWork Marketplace - Backend Dependencies

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.13.1
greenlet>=3.0.0

# Authentication
clerk-backend-api>=4.0.0
pyjwt>=2.8.0

# Payments
stripe>=7.12.0

# AI/ML
anthropic>=0.18.0
pinecone-client>=3.0.0

# Utilities
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.26.0
structlog>=24.1.0

# Redis
redis>=5.0.1

# Storage
boto3>=1.34.34

# Email
resend>=0.7.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.4
black>=24.1.1
ruff>=0.2.0
